web_scraper/data_cleaner.py
```python
import re
from logger import Logger

class DataCleaner:
    def __init__(self):
        self.logger = Logger().get_logger(__name__)

    def remove_irrelevant_info(self, data):
        try:
            # Remove any irrelevant information, advertisements, or other non-essential elements.
            cleaned_data = re.sub(r'\n\s*\n', '\n\n', data, flags=re.MULTILINE)
            return cleaned_data
        except Exception as e:
            self.logger.error(f"Error in removing irrelevant info: {str(e)}")
            return None

    def remove_duplicates(self, data_list):
        try:
            # Remove duplicates from the list of data
            cleaned_data_list = list(set(data_list))
            return cleaned_data_list
        except Exception as e:
            self.logger.error(f"Error in removing duplicates: {str(e)}")
            return None

    def clean_data(self, data):
        try:
            # Clean the data
            cleaned_data = self.remove_irrelevant_info(data)
            cleaned_data = self.remove_duplicates(cleaned_data)
            return cleaned_data
        except Exception as e:
            self.logger.error(f"Error in cleaning data: {str(e)}")
            return None
```